Move opcode fns into folder
