+++ src/gameboy/opcodes/run_fns/mod.rs
@@ -1,0 +1,36 @@
+mod add;
+mod bit;
+mod call;
+mod cp;
+mod dec;
+mod inc;
+mod jmp;
+mod ld16;
+mod ld8;
+mod or;
+mod pop;
+mod push;
+mod ret;
+mod rotate_left;
+mod rotate_left_a;
+mod rotate_method;
+mod sub;
+mod xor;
+
+pub use self::add::run_add;
+pub use self::bit::run_bit;
+pub use self::call::run_call;
+pub use self::cp::run_cp;
+pub use self::dec::run_dec;
+pub use self::inc::run_inc;
+pub use self::jmp::run_jmp;
+pub use self::ld16::run_ld16;
+pub use self::ld8::run_ld8;
+pub use self::or::run_or;
+pub use self::pop::run_pop;
+pub use self::push::run_push;
+pub use self::ret::run_ret;
+pub use self::rotate_left::run_rl;
+pub use self::rotate_left_a::run_rla;
+pub use self::sub::run_sub;
+pub use self::xor::run_xor;
