+++ src/gameboy/opcodes/run_fns/rotate_left.rs
@@ -1,16 +1,15 @@
 use crate::gameboy::cpu::CPU;
 
-use super::super::{read_flag, write_flag, Flags, RegisterLabel8};
+use super::super::super::{read_flag, write_flag, Flags};
+use super::super::Argument;
 use super::rotate_method::shift_reg_and_flag;
-use super::OpCode;
 
-impl OpCode {
-    pub fn run_rla(&self, cpu: &mut CPU, _memory: &mut Vec<u8>) -> u32 {
-        let mut cycles = 0;
-        let reg_contents = cpu.read_8_bits(RegisterLabel8::A);
+pub fn run_rl(args: &[Argument], cpu: &mut CPU, _memory: &mut Vec<u8>) -> u32 {
+    let mut cycles = 0;
+    if let Argument::Register8Constant(reg) = args[0] {
+        let reg_contents = cpu.read_8_bits(reg);
         let carry_flag = read_flag(cpu, Flags::C);
 
-        // Create the new register value
         let (new_register, new_carry) = shift_reg_and_flag(reg_contents, carry_flag);
 
         // Set the carry flag
@@ -21,14 +20,14 @@ impl OpCode {
         write_flag(cpu, Flags::N, false);
 
         // Write away the flag
-        cpu.write_8_bits(RegisterLabel8::A, new_register);
+        cpu.write_8_bits(reg, new_register);
 
         // Check the result in the C register for 0 to set the zero flag
         if new_register == 0 {
             write_flag(cpu, Flags::Z, true);
         }
 
-        cycles += 4;
-        cycles
+        cycles += 8;
     }
+    cycles
 }
