+++ src/gameboy/opcodes/run_fns/rotate_method.rs
